% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{alternative_groupings}
\alias{alternative_groupings}
\title{Use Alternative Grouping}
\usage{
alternative_groupings(file)
}
\arguments{
\item{file}{csv format file (with path) that you wish to use for Trax files groupings}
}
\value{
Nothing is returned
}
\description{
A function used for overriding the internal package data "groupings" with your own
data file (csv format). This must be run before any file conversion to allow the
alternative groupings to be used.
}
\examples{
\dontrun{alternative_groupings()}

}

