% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_broad.R
\name{changeDir}
\alias{changeDir}
\title{Change Directory}
\usage{
changeDir(dir)
}
\arguments{
\item{dir}{the name of the file path}
}
\value{
Nothing is returned
}
\description{
A simple function for changing directory
}
\examples{
changeDir(dir = "C:/user/myDocuments")

}

