% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trax_conversion.R
\name{check_pkgs}
\alias{check_pkgs}
\title{Check packages}
\usage{
check_pkgs()
}
\description{
This function checks if the local copy of R has the dependent packakges
installed. If they are available, they will be attached. If the packages are
missing, this function installs them.
}
\examples{
check_pkgs()

}

