% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_broad.R
\name{convertTraxFile}
\alias{convertTraxFile}
\title{Convert Trax File}
\usage{
convertTraxFile(filename, dir = getwd())
}
\arguments{
\item{filename}{the name along with the file path for the data}

\item{dir}{defaulted to the current working directory. This is the target
directory for your files.}
}
\value{
returns a formatted file in your given location.
}
\description{
This is function converts Trax files to the wide format for use with
Power BI.
}

