% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_broad.R
\name{createPath}
\alias{createPath}
\title{Create Converted File Name}
\usage{
createPath(filename, dir)
}
\arguments{
\item{filename}{name of original file}

\item{dir}{destination directory for converted file}
}
\value{
returns full path of directory and file name for saving
}
\description{
This is function creates a file name with the pre-fix "formatted - "
for passing to the file writing function
}

