% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groupings}
\alias{groupings}
\title{Data table for spreading the Trax files}
\format{A data frame with columns:
\describe{
 \item{Measure}{A value to join on.}
 \item{Value_Field}{New value to differentiate the rows.}
 \item{Service}{Business area of data.}
 \item{Sub_service}{Specific service within business area.}
 \item{Multiplier}{Multiplier for converting data to .}
}}
\source{
Created by Daniel Sutton.
\url{M:/All Services Visualisation/Trax files to be converted/Trax File Column Template"}
}
\usage{
groupings
}
\description{
Data table for spreading the Trax files
}
\examples{
\dontrun{
 groupings
}
}
\keyword{datasets}
