% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\docType{package}
\name{IHSTechTools}
\alias{IHSTechTools}
\alias{IHSTechTools-package}
\title{IHSTechTools: A package for all things IHS Technology}
\description{
Designed to make data ETL, analysis and visualisation easier,the
IHSTechTools package provides three categories of important functions:
Trax, Database and Visualisation.
}
\section{Trax functions}{

The Trax functions are designed for creating Trax files from the database and
manipulating Trax files for other uses e.g. Power BI.
}

\section{Database functions}{


The database functions let users access, play and push data to the MySQL database
manually.
}

\section{Visualisation functions}{


Features and functions will be added for quick visualisation.
}

