% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trax_conversion.R
\name{quarter_to_date}
\alias{quarter_to_date}
\title{Convert Quarters to dates}
\usage{
quarter_to_date(data)
}
\arguments{
\item{data}{this function takes a data frame created from a Trax file.}
}
\value{
Nothing is returned. An additional column is added to the inputted
data frame.
}
\description{
This is a function for converting the Trax quarter format to a date of the
form dd/mm/yyyy.
}

