% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{update_groupings}
\alias{update_groupings}
\title{Update Groupings data}
\usage{
update_groupings()
}
\value{
Nothing is returned
}
\description{
A function used for editing the internal package data "groupings".
When run, you will be prompted to give the file which you wish to
replace the internal data with. It will then be added to the package.
This function will only run if you are in the package.
}
\examples{
\dontrun{update_groupings()}

}
