% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trax_conversion.R
\name{changeDir}
\alias{changeDir}
\title{Change Directory}
\usage{

  changeDir(dir = "M:/All Services Visualisation/Trax files to be converted/Files")
}
\arguments{
\item{dir}{the name of the file path}
}
\value{
Nothing is returned
}
\description{
A simple function for changing directory.
I am adding a test text line
}
\examples{
changeDir(dir = getwd())

}
