% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trax_conversion.R
\name{convertAll}
\alias{convertAll}
\title{Convert All Trax Files}
\usage{
convertAll(dir = getwd(), dest)
}
\arguments{
\item{dir}{the name of the file path}

\item{dest}{The path where your converted files will be saved.}
}
\value{
returns a formatted version of all the files in the directory
}
\description{
This is function file handler is for converting all files in the
given directory.
}

